Minor insight tweak
