+++ src/launchpad/insights/common.py
@@ -3,57 +3,16 @@
 from __future__ import annotations
 
 from collections import defaultdict
-from dataclasses import dataclass
-from typing import Dict, List, Protocol, TypeVar
+from typing import Dict, List
 
-from launchpad.models.common import FileAnalysis, FileInfo
+from launchpad.insights.insight import Insight, InsightsInput
+from launchpad.models.common import FileInfo
 from launchpad.models.insights import DuplicateFilesInsightResult
-from launchpad.models.treemap import TreemapResults
-
-from ..models.apple import AppleAppInfo, MachOBinaryAnalysis
-
-T_co = TypeVar("T_co", covariant=True)
-
-
-@dataclass
-class InsightsInput:
-    app_info: AppleAppInfo
-    file_analysis: FileAnalysis
-    treemap: TreemapResults | None
-    binary_analysis: List[MachOBinaryAnalysis]
-
-
-class Insight(Protocol[T_co]):
-    """Protocol for insight functions.
-
-    Insights are functions that take analysis results and return typed insight results.
-    All data needed for the insight must be collected during the main analysis phase.
-    """
-
-    def __call__(self, input: InsightsInput) -> T_co:
-        """Generate insights from analysis results.
-
-        Args:
-            results: The analysis results to generate insights from
-
-        Returns:
-            Typed insight results
-        """
-        ...
 
 
 class DuplicateFilesInsight(Insight[DuplicateFilesInsightResult]):
-    """Insight for duplicate files analysis."""
 
-    def __call__(self, input: InsightsInput) -> DuplicateFilesInsightResult:
-        """Generate insights about duplicate files.
-
-        Args:
-            results: The analysis results to generate insights from
-
-        Returns:
-            Duplicate files insight results
-        """
+    def get_results(self, input: InsightsInput) -> DuplicateFilesInsightResult:
         # Group files by hash
         files_by_hash: Dict[str, List[FileInfo]] = defaultdict(list)
         for file in input.file_analysis.files:
